ORC-1200: Extracting encryption setup logic from `WriterImpl`.
